requirements.txt
pygame
pyinstaller
pymongo
rasterio
tqdm
python-dotenv